\name{jnt-package}
\alias{jnt-package}
\alias{jnt-package}
\title{Johnson-Neyman technique...}
\description{Johnson-Neyman technique
Carry out the Johnson-Neyman Technique}
\details{\tabular{ll}{ Package: \tab jnt\cr Type: \tab Package\cr Version: \tab
0.1-2\cr Date: \tab 2011-01-30\cr License: \tab GPL-2\cr LazyLoad: \tab
yes\cr LazyData: \tab yes\cr }}
\docType{package}
\author{Kevin Middleton (kmm@csusb.edu)

Maintainer: Kevin Middleton (kmm@csusb.edu)}
\references{Johnson PO, Neyman J (1936) Tests of certain linear hypotheses
and their application to some educational problems. \emph{Statistical
Research Memoirs} 1:57-93.

Hunka S, Leighton J (1997) Defining Johnson-Neyman regions of significance
in three-covariate ANCOVA using mathematica. \emph{Journal of Educational
and Behavioral Statistics} 22: 361-387.

White CR (2003) Allometric analysis beyond heterogenous regression slopes:
Use of the Johnson-Neyman Technique in comparative biology. \emph{Physiol
Biochem Zool} 76: 135-140.

\emph{Examples:}

White CR (2003) The influence of foraging mode and arid adaptation on the
basal metabolic rates of burrowing mammals. \emph{Physiol Biochem Zool}
76: 122-134.

Lavin SR, Karasov WH, Ives AR, Middleton KM, Garland T (2008) Morphometrics
of the avian small intestine compared with that of nonflying mammals: A
phylogenetic approach. \emph{Physiol Biochem Zool} 81: 526-550.}
\keyword{package}
\examples{example(jnt)}

