\name{Lavin.2}
\alias{Lavin.2}
\alias{Lavin.2}
\title{Data from Lavin et al.}
\description{Data from Lavin et al., 2008}
\details{Data set 2 from Lavin et al., 2008}
\docType{data}
\references{Lavin SR, Karasov WH, Ives AR, Middleton KM, Garland T, Jr (2008)
Morphometrics of the avian small intestine compared with that of nonflying
mammals: A phylogenetic approach. \emph{Physiol Biochem Zool} 81: 526-550.}
\keyword{datasets}
\examples{str(Lavin.1)
str(Lavin.2)

(Lavin <- jnt(Lavin.1, Lavin.2))
plot(Lavin)}

