% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.jnt.R
\name{plot.jnt}
\alias{plot.jnt}
\title{Plot jnt Object}
\usage{
\method{plot}{jnt}(x, reg_lty = 1, jnt_lty = 2, col_signif = "red",
  plot_range = TRUE, ...)
}
\arguments{
\item{x}{Object of class \code{jnt}}

\item{reg_lty}{Line type for regression lines. Defaults to 1.}

\item{jnt_lty}{Line type for range of insignificant differences.
Defaults to 2.}

\item{col_signif}{Color for bars denoting range of insignificance.
Defaults to "red."}

\item{plot_range}{logical. If \code{TRUE} (the default) the range
of insignificance is plotted.}

\item{...}{Additional parameters passed to ggplot (e.g., lwd, pch,
xlab, ylab).}
}
\description{
Plot jnt Object
}
\details{
Plot object returned by \code{jnt}.
}
\examples{

example(jnt)

}
\author{
Kevin Middleton \email{middletonk@missouri.edu}
}
\keyword{univar}

