\name{White.1}
\alias{White.1}
\alias{White.1}
\title{Data from White, 2003...}
\description{Data from White, 2003}
\details{Data set 1 from from White, 2003}
\docType{data}
\references{White CR (2003) Allometric analysis beyond heterogenous regression
slopes: Use of the Johnson-Neyman Technique in comparative biology.
\emph{Physiol Biochem Zool} 76: 135-140.}
\keyword{datasets}
\examples{str(White.1)
str(White.2)

(White <- jnt(White.1, White.2))
plot(White)}

