\name{plot.jnt}
\alias{plot.jnt}
\title{Plot jnt Object...}
\usage{\method{plot}{jnt}(x, reg.lty=1, jnt.lty=2, plot.range=TRUE, ...)
}
\description{Plot jnt Object}
\details{Plot object returned by \code{jnt}.}
\author{Kevin Middleton (kmm@csusb.edu)}
\keyword{univar}
\arguments{\item{x}{Object of class \code{jnt}}
\item{reg.lty}{Line type for regression lines. Defaults to 1.}
\item{jnt.lty}{Line type for range of significance. Defaults to 2.}
\item{plot.range}{logical. If \code{TRUE} (the default) the range of significance is plotted.}
\item{...}{Additional parameters passed to plot}
}
\examples{example(jnt)}

